Change: move functions to ssh.rs.
Ssh functions are not implemented methods for a session object anymore.
