+++ rust/nasl-interpreter/Cargo.toml
@@ -29,4 +29,4 @@ ctr = "0.9.2"
 aes-gcm = { version = "0.10.1"}
 cmac = "0.7.2"
 #libssh-rs = "0.1.8"
-libssh-rs = {git = "https://github.com/jjnicola/libssh-rs.git", branch = "ssh-read"}
+libssh-rs = {git = "https://github.com/wez/libssh-rs.git", branch = "main"}
